Create replace.java
next() and nextLine() methods are very useful in the String.

next() will be stop if the line of String have Spaces. [This method can read the input only until a space(” “) is encountered. In other words, it finds and returns the next complete token from the scanner. ]
nextLine()-->[This method can read the input till the end of line. In other words, it can take input until the line change or new line and ends input of getting ‘\n’ or press enter. ]